refactor(docs-retrieval): extract package + resolver to package-fetcher.ts (Phase 5)
Move the package-resolver singleton and its 9 consumer functions out of
content-fetcher.ts into a dedicated package-fetcher.ts module. Pattern J
contract-surface extraction; Pattern H (resolver injection) explicitly
deferred per the investigation.

Symbols moved (atomic, no behavior change):
  - _packageResolver       (module-scope `let` — the singleton)
  - setPackageResolver
  - buildMilestoneWebsiteUrl
  - derivePathSlug
  - resolvePackageMilestones
  - resolvePackageNavLinks
  - isPathManifest
  - getManifestMilestoneIds
  - fetchPackageContent
  - fetchPackageById

Symbols that STAY in content-fetcher.ts:
  - injectJourneyExtrasIntoJsonGuide (called by docs-panel directly and
    by fetchPackageContent via cross-module import; staying put per PLAN
    decision DR-04). package-fetcher.ts imports it explicitly.

Public API of docs-retrieval is unchanged: `index.ts` barrel re-exports
the same six symbols (fetchPackageContent, fetchPackageById,
setPackageResolver, resolvePackageMilestones, resolvePackageNavLinks,
derivePathSlug) with the source path swapped from `./content-fetcher`
to `./package-fetcher`. External callers see no change.

Two in-tree test files are ownership-path updated to point at the new
module (permitted exception per Pattern J):
  - package-fetcher.pre-extraction.test.ts -> package-fetcher.test.ts
    (rename via git mv; imports rewritten to ./package-fetcher)
  - package-content.test.ts: split import — package-fetcher symbols
    from ./package-fetcher, fetchContent stays on ./content-fetcher

content-fetcher.ts loses the now-unused type imports
(PackageResolver, ResolvedNavLink, Milestone, getPackageRenderType).

Test plan:
  - npm run typecheck     : clean
  - npm run lint           : clean
  - npm run prettier       : clean
  - npm run test:ci        : 2539 passed / 18 skipped / 131 suites
                             (was 2520 / 18 / 130 → +19 promoted, +1 suite)
  - npm run build          : clean (4 pre-existing warnings only)

Refs: .refactor/PLAN.md Phase 5 (Pattern J + deferred H, HIGH)
